Minor markup cleanup
