+++ locales/en/translation.json
@@ -263,6 +263,7 @@
   "targetDirectoryPreInfo": "Pre-Install Directories are for unmanaged metadata components that should always be deployed prior to deploying or installing the package.",
   "targetDirectorySourceInfo": "Package Directories hold the main metadata components of the package you are building. These components will be deployed as an unmanaged package during development, but installed as a managed or unlocked package in production.",
   "taskNotFound": "We can’t find the task you’re looking for. Try <2>another task</2> from that project?",
+  "tasksComplete": "<0>{{complete}}</0> of <2>{{total}}</2> Complete",
   "up to date": "up to date",
   "view changes": "view changes",
   "viewOutdatedOrg": "You may proceed with the outdated org, but be aware that you will not be reviewing the latest changes.",
