+++ src/js/store/projects/reducer.ts
@@ -15,11 +15,13 @@ export interface Project {
   branch_diff_url: string | null;
   pr_url: string | null;
   pr_is_open: boolean;
+  pr_is_merged: boolean;
   has_unmerged_commits: boolean;
   currently_creating_pr: boolean;
   github_users: GitHubUser[];
   status: ProjectStatuses;
 }
+
 export interface ProjectsByRepositoryState {
   projects: Project[];
   next: string | null;
